Fixed #53: added --exclude option
Exclude takes a path or URL to a text file containing a list of articles or categories.
Items in the list must be Article ID (path part of URL) or Category ID using the `Category:` prefix.
Blank and `#` starting lines are ignored.

Items doesn't have to exist on the target wikiHow. It's possible to manage a single list for
all languages for instance.

The exclusion is done while walking down the categories list so the scraper does not
enter an excluded category and thus doesn't scrape its articles.

Same goes for articles. When looking for articles listed in a category, those that are
to be excluded are not scraped.

In addition, the scraper tries to remove links to those excluded Categories and Articles
so that if you are excluding an article in a non-excluded category, the article is not
listed on the category page.

- breadcrumbs is rewriten to remove links to excluded categories
- links removal is done in rewriter, so applies to all pages

Consequences & limits:

- Category page usually contains a defined number of article links so that the grid
of links looks complete. When excluding articles, we remove links to them here, shifting
the cells. Doesn't look broken, but can be un-even.
- Same applies to related articles link at the bottom and on the sidebar of articles.
- Breadcrumb (top and footer) would look different if a category is excluded of course.
- We look for identified uses of article and category links to remove the whole block
so that you don't get a thumbnail to an unwanted content. There might be non-identified
links as well. For those, we keep the element but remove the target (href).
- While exclusion is recursive, links rewriting is not. If you exclude Category:Sex
and inside that category is an article named Having-Sex, then the article is excluded
and can't be accessed from that category (as it is not included). But if an article in
Category:Potato-Chips has an article that lists Having-sex as a related article, then
there would be a link to Having-Sex… that would lead to a 404. Not really avoidable
as we don't have a graph of categories and articles.
I don't know how often those situation might happen but I suppose for the intended usecase
of filtering-out sexual-related material, there is probably few cross-posting.
It's possible to add such outlier articles to the exclude list to remove those links.