Speed-up video processing when using --without-videos
--without-videos only prevent the Youtube ones from being processed and included
but there are many regular videos anyway.
Those regular videos are short and small-sized and easy to process.

Youtube ones can be huge and are retrieved from Youtube where exceeding a certain
number of resource usage (requests? bandwith ?) results in being banned for a period
of time.

We thus need to process Youtube videos one by one but others can be handled concurrently.

This changes the nb or workers for videos based on --without-videos.

This is a simple fix for the most common use case (--without-videos) but ideally,
we should treat both sources differently no matter the option as including Youtube
videos means processing __all__ the videos on a single worker.