Fixed #94: redirects on categories
In Thai wikiHow, many categories are reachable via redirections with different names.
This behavior has not been seen on other languages.

Updated scraper behavior:
- start scraping from CategoryListing instead of Sitemap. This is the page that we
use as homepage so it makes more sense to use it to start crawling.
See [1]
- Assume category pages can be redirections. We still follow redirections when fetching
content but now we are also keeping track of redirections and adding a ZIM redirection
for each of the redirection we encountered while building a category page

[1] In case of Thai, Sitemap has a link to the actual category page, without redirection.
CategoryListing has a different link that uses a redirection.
If using Sitemap for crawling, we don't see the redirect link and thus, as we use
CategoryLink for homepage, we get a 404.
Using CategoryListing ensures what the users sees is what we crawled from.

Example:
Sitemap has link to:            หมวดหมู่:การดูแลและไลฟ์สไตล์
That resolves at:               หมวดหมู่:การดูแลและไลฟ์สไตล์

Categorylisting has link to:    หมวดหมู่:การดูแลตัวเองและรสนิยมส่วนตัว
That resolves at:               หมวดหมู่:การดูแลและไลฟ์สไตล์