Fixed #103: single --category fixes homepage
When a single --category param is passed, it is assumed the run is for a single
category. In this case, the category is displayed as Home page:

/Home and /Special:CategoryListing are redirected to it as well

WH API is queried to get list of categories used in this category (should this category
be a sub-category) and those are added to the excluded categories list so that
links rewriting removes them from breadcrumbs.