Fixed #110: added --only filter
To be used in conjonction with a single category, an --only param allows pointing to
a list of Articles to include only.

All found articles, scraping that category are only included if in the list.

This “positive list” allows removing related links in articles that are not in the list.