Fixed #112: more solid categories retrieval
In single category mode, we fetch the categories attached to the requested category
and add them to the exlcuded_categories list to prevent adding parent categories.

Because of how the API works, an incomplete request for categories because part of
the category was removed (considered a different param) still returns categories
matching requested name… and this the required category.
The category ends up being added to the exclusion list and there is nothing to scrape.

This fixes the API call request encoding.